backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1

# Machine Learning
tensorflow==2.15.0
torch==2.1.1
torchvision==0.16.1
scikit-learn==1.3.2
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.24.3
pandas==2.1.4

# Image Processing
Pillow==10.1.0
opencv-python==4.8.1.78

# HTTP Requests
httpx==0.25.2
requests==2.31.0

# Environment and Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# CORS and Security
python-cors==1.7.0
cryptography==41.0.8

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Monitoring and Logging
structlog==23.2.0
